requirements.txt
discord.py
flask
requests
python-dotenv
cohere
psycopg[binary]
